Refactor aggregation tests and enhance query executor
- Updated `groupCounts` type in `AggregationOperationTests` to `long` for consistency.
- Bumped version to `1.0.0-preview.8` in the project file.
- Added support for `.by()` modulator steps in `TinkerGraphQueryExecutor`.
- Improved predicate handling in `EvaluatePredicate` for better query flexibility.
- Refactored `ExecuteGroupStep` to support dual `.by()` modulators.
- Introduced new test classes `DebugChainedTest` and `GroupDebugTest` for feature validation.
- Enhanced `GetCurrentResults` method for automatic deduplication of results.
- Added various methods to improve traversal context handling.

+++ src/Stardust.Paradox.Data.InMemory.Tests/AggregationOperationTests.cs
@@ -141,7 +141,7 @@ public class AggregationOperationTests
 
         // Assert
         result.Should().HaveCount(1);
-        var groupCounts = result.First() as Dictionary<string, int>;
+        var groupCounts = result.First() as Dictionary<string, long>;
         groupCounts.Should().NotBeNull();
     }
 
@@ -336,7 +336,7 @@ public class AggregationOperationTests
 
         // Assert
         result.Should().HaveCount(1);
-        var groupCounts = result.First() as Dictionary<string, int>;
+        var groupCounts = result.First() as Dictionary<string, long>;
         groupCounts.Should().NotBeNull();
         groupCounts.Should().HaveCountGreaterThan(0);
     }
@@ -485,12 +485,12 @@ public class AggregationOperationTests
 
         // Assert
         result.Should().HaveCount(1);
-        var groupCounts = result.First() as Dictionary<string, int>;
+        var groupCounts = result.First() as Dictionary<string, long>; // Changed from int to long for consistency
         groupCounts.Should().NotBeNull();
         groupCounts.Should().ContainKey("IT");
         groupCounts.Should().ContainKey("HR");
-        groupCounts["IT"].Should().Be(2);
-        groupCounts["HR"].Should().Be(1);
+        groupCounts["IT"].Should().Be(2L); // Changed from 2 to 2L
+        groupCounts["HR"].Should().Be(1L); // Changed from 1 to 1L
     }
 
     [Fact]
